add middelware add auth

+++ resources/views/admin/post/list.blade.php
@@ -5,21 +5,49 @@
 @section('main')
 <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-3 pb-2 mb-3 border-bottom">
   <h1 class="h2">Post / List </h1>
+        <?php if ($auth_add): ?>
   <div class="btn-toolbar mb-2 mb-md-0">
     <a href="/admin/post/add" class="btn btn-outline-secondary">Add Post</a>
   </div>
+        <?php endif ?>
 </div>
 
 
 <?php if (count($list)): ?>
   <table class="table table-striped">
+    <thead>
+      <tr>
+        <td>Title</td>
+        <td>Published</td>
+        <td>action</td>
+      </tr>
+    </thead>
     <tbody>
       <?php foreach ($list as $key => $value): ?>
       <tr>
         <td>{{$value->title}}</td>
         <td>
+            
+          <input type="checkbox" class="form-check-input published" data-post-id="{{ $value->post_id }}" name="published"
+          <?php if (!$auth_publish): ?>
+            disabled
+          <?php endif ?>
+          <?php if ($value->publish): ?>
+            checked
+          <?php endif ?>
+          >
+        </td>
+        <td>
+
+        <?php if ($auth_edit): ?>
           <a href="/admin/post/edit?post_id={{ $value->post_id }}" class="btn btn-primary">Edit</a>
+
+        <?php endif ?>
+
+        <?php if ($auth_delete): ?>
+
           <a href="/admin/post/delete?post_id={{ $value->post_id }}" class="btn btn-danger">Delete</a>
+        <?php endif ?>
         </td>
       </tr>
       <?php endforeach ?>
@@ -30,5 +58,14 @@
 <?php endif ?>
 
 
-
+<script type="text/javascript">
+  $('.published').change(function () {
+    data = {post_id:$(this).attr('data-post-id'),value:$(this).prop('checked')}
+    $.ajax({
+      type:"GET",
+      url:"/admin/post/publish",
+      data:data,
+    })
+  });
+</script>
 @endsection
